Templated cross product function

+++ src/efaroe.h
@@ -5,41 +5,42 @@
 #include <iostream>
 #include <math.h>
 
+#include "math_util.h"
 #include "quaternions.h"
 
 class eFaroe
 {
     public:
-        eFaroe(quaternion initial_quaternion, std::array<float, 3> gyro_bias, float gyro_error, int use_mag);
-        void update(unsigned long timestamp, float ax, float ay, float az, float gx, float gy, float gz);
+        eFaroe(quaternion initial_quaternion, std::array<double, 3> gyro_bias, double gyro_error, int use_mag);
+        void update(unsigned long timestamp, double ax, double ay, double az, double gx, double gy, double gz);
 
-        float getW();
-        float getX();
-        float getY();
-        float getZ();
+        double getW();
+        double getX();
+        double getY();
+        double getZ();
 
-        float getPitch();
-        float getRoll();
-        float getYaw();
+        double getPitch();
+        double getRoll();
+        double getYaw();
 
         quaternion getQuat();
-        std::array<float, 3> getEuler();
+        std::array<double, 3> getEuler();
     
     private:
-        std::array<float, 3> gyro_bias;
-        float gyro_error;
-        float true_error;
-        float uk_dip;
-        std::array<float, 2> emf;
-        float zeta;
-        float beta;
+        std::array<double, 3> gyro_bias;
+        double gyro_error;
+        double true_error;
+        double uk_dip;
+        std::array<double, 2> emf;
+        double zeta;
+        double beta;
         unsigned long last_read;
         quaternion initial_quaternion{1, 0, 0, 0};
         unsigned int conv_count;
 
-        float pitch;
-        float roll;
-        float yaw;
+        double pitch;
+        double roll;
+        double yaw;
 
         void update_mag_imu();
         void update_imu();
